requirements.txt
click==7.1.2
exif==1.2.1
ImageHash==4.2.0
numpy==1.20.2
pika==1.2.0
Pillow==8.2.0
plum-py==0.3.1
py3exiv2==0.9.3
python-dotenv==0.16.0
PyWavelets==1.1.1
scipy==1.6.2
six==1.15.0
